+++ tensorrt_llm/commands/serve.py
@@ -229,6 +229,123 @@ def launch_server(
         asyncio.run(server(host, port, sockets=[s]))
 
 
+def launch_grpc_server(host: str, port: int, llm_args: dict):
+    """
+    Launch a gRPC server for TensorRT-LLM.
+
+    This provides a high-performance gRPC interface designed for external routers
+    (e.g., sgl-router) using pre-tokenized input and raw token ID output.
+
+    Args:
+        host: Host to bind to
+        port: Port to bind to
+        llm_args: Arguments for LLM initialization (from get_llm_args)
+    """
+    import grpc
+
+    try:
+        from grpc_reflection.v1alpha import reflection
+        REFLECTION_AVAILABLE = True
+    except ImportError:
+        REFLECTION_AVAILABLE = False
+
+    from tensorrt_llm.grpc import trtllm_engine_pb2, trtllm_engine_pb2_grpc
+    from tensorrt_llm.grpc.grpc_request_manager import GrpcRequestManager
+    from tensorrt_llm.grpc.grpc_servicer import TrtLlmEngineServicer
+
+    async def serve_grpc_async():
+        logger.info("Initializing TensorRT-LLM gRPC server...")
+
+        backend = llm_args.get("backend")
+        model_path = llm_args.get("model", "")
+
+        if backend == "pytorch":
+            llm_args.pop("build_config", None)
+            llm = PyTorchLLM(**llm_args)
+        elif backend == "_autodeploy":
+            from tensorrt_llm._torch.auto_deploy import LLM as AutoDeployLLM
+            llm_args.pop("build_config", None)
+            llm = AutoDeployLLM(**llm_args)
+        elif backend == "tensorrt" or backend == "trt":
+            llm_args.pop("backend")
+            llm = LLM(**llm_args)
+        else:
+            raise click.BadParameter(
+                f"{backend} is not a known backend, check help for available options.",
+                param_hint="backend")
+
+        logger.info("Model loaded successfully")
+
+        # Create request manager
+        request_manager = GrpcRequestManager(llm)
+
+        # Create servicer
+        servicer = TrtLlmEngineServicer(request_manager, model_path=model_path)
+
+        # Create gRPC server
+        server = grpc.aio.server(
+            options=[
+                ("grpc.max_send_message_length", -1),  # Unlimited
+                ("grpc.max_receive_message_length", -1),  # Unlimited
+                ("grpc.keepalive_time_ms", 30000),  # 30s keepalive
+                ("grpc.keepalive_timeout_ms", 10000),  # 10s timeout
+            ],
+        )
+
+        # Add servicer to server
+        trtllm_engine_pb2_grpc.add_TrtLlmEngineServicer_to_server(servicer, server)
+
+        # Enable reflection for grpcurl and other tools
+        if REFLECTION_AVAILABLE:
+            service_names = (
+                trtllm_engine_pb2.DESCRIPTOR.services_by_name["TrtLlmEngine"].full_name,
+                reflection.SERVICE_NAME,
+            )
+            reflection.enable_server_reflection(service_names, server)
+            logger.info("gRPC reflection enabled")
+
+        # Bind to address
+        address = f"{host}:{port}"
+        server.add_insecure_port(address)
+
+        # Start server
+        await server.start()
+        logger.info(f"TensorRT-LLM gRPC server started on {address}")
+        logger.info("Server is ready to accept requests")
+
+        # Handle shutdown signals
+        loop = asyncio.get_running_loop()
+        stop_event = asyncio.Event()
+
+        def signal_handler():
+            logger.info("Received shutdown signal")
+            stop_event.set()
+
+        for sig in (signal.SIGTERM, signal.SIGINT):
+            loop.add_signal_handler(sig, signal_handler)
+
+        # Serve until shutdown signal
+        try:
+            await stop_event.wait()
+        except KeyboardInterrupt:
+            logger.info("Interrupted by user")
+        finally:
+            logger.info("Shutting down TensorRT-LLM gRPC server...")
+
+            # Stop gRPC server
+            await server.stop(grace=5.0)
+            logger.info("gRPC server stopped")
+
+            # Shutdown LLM
+            if hasattr(llm, "shutdown"):
+                llm.shutdown()
+            logger.info("LLM engine stopped")
+
+            logger.info("Shutdown complete")
+
+    asyncio.run(serve_grpc_async())
+
+
 def launch_mm_encoder_server(
     host: str,
     port: int,
@@ -643,6 +760,131 @@ def serve_encoder(model: str, host: str, port: int, log_level: str,
     launch_mm_encoder_server(host, port, encoder_args, metadata_server_cfg)
 
 
+@click.command("grpc")
+@click.argument("model", type=str)
+@click.option("--tokenizer",
+              type=str,
+              default=None,
+              help="Path | Name of the tokenizer.")
+@click.option("--host",
+              type=str,
+              default="0.0.0.0",
+              help="Hostname of the gRPC server.")
+@click.option("--port", type=int, default=50051, help="Port of the gRPC server.")
+@click.option(
+    "--backend",
+    type=ChoiceWithAlias(["pytorch", "tensorrt", "_autodeploy"],
+                         {"trt": "tensorrt"}),
+    default="pytorch",
+    help="The backend to use to serve the model. Default is pytorch backend.")
+@click.option('--log_level',
+              type=click.Choice(severity_map.keys()),
+              default='info',
+              help="The logging level.")
+@click.option("--max_beam_width",
+              type=int,
+              default=BuildConfig.model_fields["max_beam_width"].default,
+              help="Maximum number of beams for beam search decoding.")
+@click.option("--max_batch_size",
+              type=int,
+              default=BuildConfig.model_fields["max_batch_size"].default,
+              help="Maximum number of requests that the engine can schedule.")
+@click.option(
+    "--max_num_tokens",
+    type=int,
+    default=BuildConfig.model_fields["max_num_tokens"].default,
+    help=
+    "Maximum number of batched input tokens after padding is removed in each batch."
+)
+@click.option(
+    "--max_seq_len",
+    type=int,
+    default=BuildConfig.model_fields["max_seq_len"].default,
+    help="Maximum total length of one request, including prompt and outputs. "
+    "If unspecified, the value is deduced from the model config.")
+@click.option("--tp_size", type=int, default=1, help='Tensor parallelism size.')
+@click.option("--pp_size",
+              type=int,
+              default=1,
+              help='Pipeline parallelism size.')
+@click.option("--ep_size",
+              type=int,
+              default=None,
+              help="expert parallelism size")
+@click.option("--cluster_size",
+              type=int,
+              default=None,
+              help="expert cluster parallelism size")
+@click.option("--gpus_per_node",
+              type=int,
+              default=None,
+              help="Number of GPUs per node. Default to None, and it will be "
+              "detected automatically.")
+@click.option("--kv_cache_free_gpu_memory_fraction",
+              type=float,
+              default=0.9,
+              help="Free GPU memory fraction reserved for KV Cache, "
+              "after allocating model weights and buffers.")
+@click.option("--trust_remote_code",
+              is_flag=True,
+              default=False,
+              help="Flag for HF transformers.")
+@click.option(
+    "--extra_llm_api_options",
+    type=str,
+    default=None,
+    help=
+    "Path to a YAML file that overwrites the parameters specified by trtllm-serve."
+)
+@click.option(
+    "--reasoning_parser",
+    type=click.Choice(ReasoningParserFactory.parsers.keys()),
+    default=None,
+    help="[Experimental] Specify the parser for reasoning models.",
+)
+def grpc(
+        model: str, tokenizer: Optional[str], host: str, port: int,
+        log_level: str, backend: str, max_beam_width: int, max_batch_size: int,
+        max_num_tokens: int, max_seq_len: int, tp_size: int, pp_size: int,
+        ep_size: Optional[int], cluster_size: Optional[int],
+        gpus_per_node: Optional[int], kv_cache_free_gpu_memory_fraction: float,
+        trust_remote_code: bool, extra_llm_api_options: Optional[str],
+        reasoning_parser: Optional[str]):
+    """Running a gRPC server for high-performance communication
+
+    MODEL: model name | HF checkpoint path | TensorRT engine path
+
+    The gRPC server accepts pre-tokenized requests and returns raw token IDs,
+    enabling efficient binary communication with external routers like sgl-router.
+    """
+    logger.set_level(log_level)
+
+    llm_args, _ = get_llm_args(
+        model=model,
+        tokenizer=tokenizer,
+        backend=backend,
+        max_beam_width=max_beam_width,
+        max_batch_size=max_batch_size,
+        max_num_tokens=max_num_tokens,
+        max_seq_len=max_seq_len,
+        tensor_parallel_size=tp_size,
+        pipeline_parallel_size=pp_size,
+        moe_expert_parallel_size=ep_size,
+        moe_cluster_parallel_size=cluster_size,
+        gpus_per_node=gpus_per_node,
+        free_gpu_memory_fraction=kv_cache_free_gpu_memory_fraction,
+        trust_remote_code=trust_remote_code,
+        reasoning_parser=reasoning_parser)
+
+    llm_args_extra_dict = {}
+    if extra_llm_api_options is not None:
+        with open(extra_llm_api_options, 'r') as f:
+            llm_args_extra_dict = yaml.safe_load(f)
+    llm_args = update_llm_args_with_extra_dict(llm_args, llm_args_extra_dict)
+
+    launch_grpc_server(host, port, llm_args)
+
+
 @click.command("disaggregated")
 @click.option("-c",
               "--config_file",
@@ -959,6 +1201,7 @@ class DefaultGroup(click.Group):
 main = DefaultGroup(
     commands={
         "serve": serve,
+        "grpc": grpc,
         "disaggregated": disaggregated,
         "disaggregated_mpi_worker": disaggregated_mpi_worker,
         "mm_embedding_serve": serve_encoder
