+++ tensorrt_llm/grpc/grpc_servicer.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python3
 # SPDX-FileCopyrightText: Copyright (c) 2024 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 #
@@ -15,67 +14,28 @@
 # limitations under the License.
 
 """
-TensorRT-LLM gRPC Server
+gRPC Servicer for TensorRT-LLM
 
-Starts a gRPC server for TensorRT-LLM using the TrtLlmEngine protocol.
-Designed for high-performance communication with external routers (e.g., sgl-router)
-using pre-tokenized input and raw token ID output.
-
-This server inherits all arguments from the standard TensorRT-LLM serve command,
-ensuring consistent behavior with the OpenAI-compatible HTTP server.
-
-Key Features:
-- Pre-tokenized input (no Python tokenization overhead)
-- Raw token ID output (no Python detokenization overhead)
-- Streaming support with delta tokens
-- Full sampling parameter support
-- Guided decoding (JSON schema, regex, grammar)
-- LoRA and prompt tuning support
-- Disaggregated inference support
-
-Usage:
-    python -m tensorrt_llm.entrypoints.grpc_server <model_path> [OPTIONS]
-
-Example:
-    python -m tensorrt_llm.entrypoints.grpc_server \
-        /raid/models/meta-llama/Llama-3.1-8B-Instruct \
-        --backend pytorch \
-        --host 0.0.0.0 \
-        --port 50051 \
-        --tp_size 1 \
-        --trust_remote_code
+Implements the TrtLlmEngine gRPC service for high-performance communication
+with external routers (e.g., sgl-router) using pre-tokenized input.
 """
 
 import asyncio
-import signal
 import time
 from collections.abc import AsyncGenerator
 from typing import Optional
 
-import click
 import grpc
-import yaml
 
-try:
-    from grpc_reflection.v1alpha import reflection
-    REFLECTION_AVAILABLE = True
-except ImportError:
-    REFLECTION_AVAILABLE = False
+from tensorrt_llm.logger import logger
 
-from tensorrt_llm import LLM as PyTorchLLM
-from tensorrt_llm._tensorrt_engine import LLM
-from tensorrt_llm.commands.serve import get_llm_args
-from tensorrt_llm.grpc import trtllm_engine_pb2, trtllm_engine_pb2_grpc
-from tensorrt_llm.grpc.grpc_request_manager import (
+from . import trtllm_engine_pb2, trtllm_engine_pb2_grpc
+from .grpc_request_manager import (
     GrpcRequestManager,
     create_sampling_params_from_proto,
     create_lora_request_from_proto,
     create_disaggregated_params_from_proto,
 )
-from tensorrt_llm.llmapi import BuildConfig
-from tensorrt_llm.llmapi.llm_utils import update_llm_args_with_extra_dict
-from tensorrt_llm.llmapi.reasoning_parser import ReasoningParserFactory
-from tensorrt_llm.logger import logger, severity_map
 
 
 class TrtLlmEngineServicer(trtllm_engine_pb2_grpc.TrtLlmEngineServicer):
@@ -483,254 +443,3 @@ class TrtLlmEngineServicer(trtllm_engine_pb2_grpc.TrtLlmEngineServicer):
                 code=code,
             ),
         )
-
-
-async def serve_grpc(
-    host: str,
-    port: int,
-    llm_args: dict,
-    model_path: str,
-):
-    """
-    Main serving function.
-
-    Args:
-        host: Host to bind to
-        port: Port to bind to
-        llm_args: Arguments for LLM initialization (from get_llm_args)
-        model_path: Path to the model
-    """
-    logger.info("Initializing TensorRT-LLM gRPC server...")
-
-    # Create LLM based on backend
-    backend = llm_args.get("backend")
-    if backend == "pytorch":
-        llm = PyTorchLLM(**llm_args)
-    else:
-        llm = LLM(**llm_args)
-
-    logger.info("Model loaded successfully")
-
-    # Create request manager
-    request_manager = GrpcRequestManager(llm)
-
-    # Create servicer
-    servicer = TrtLlmEngineServicer(request_manager, model_path=model_path)
-
-    # Create gRPC server
-    server = grpc.aio.server(
-        options=[
-            ("grpc.max_send_message_length", -1),  # Unlimited
-            ("grpc.max_receive_message_length", -1),  # Unlimited
-            ("grpc.keepalive_time_ms", 30000),  # 30s keepalive
-            ("grpc.keepalive_timeout_ms", 10000),  # 10s timeout
-        ],
-    )
-
-    # Add servicer to server
-    trtllm_engine_pb2_grpc.add_TrtLlmEngineServicer_to_server(servicer, server)
-
-    # Enable reflection for grpcurl and other tools
-    if REFLECTION_AVAILABLE:
-        service_names = (
-            trtllm_engine_pb2.DESCRIPTOR.services_by_name["TrtLlmEngine"].full_name,
-            reflection.SERVICE_NAME,
-        )
-        reflection.enable_server_reflection(service_names, server)
-        logger.info("gRPC reflection enabled")
-
-    # Bind to address
-    address = f"{host}:{port}"
-    server.add_insecure_port(address)
-
-    # Start server
-    await server.start()
-    logger.info(f"TensorRT-LLM gRPC server started on {address}")
-    logger.info("Server is ready to accept requests")
-
-    # Handle shutdown signals
-    loop = asyncio.get_running_loop()
-    stop_event = asyncio.Event()
-
-    def signal_handler():
-        logger.info("Received shutdown signal")
-        stop_event.set()
-
-    for sig in (signal.SIGTERM, signal.SIGINT):
-        loop.add_signal_handler(sig, signal_handler)
-
-    # Serve until shutdown signal
-    try:
-        await stop_event.wait()
-    except KeyboardInterrupt:
-        logger.info("Interrupted by user")
-    finally:
-        logger.info("Shutting down TensorRT-LLM gRPC server...")
-
-        # Stop gRPC server
-        await server.stop(grace=5.0)
-        logger.info("gRPC server stopped")
-
-        # Shutdown LLM
-        if hasattr(llm, "shutdown"):
-            llm.shutdown()
-        logger.info("LLM engine stopped")
-
-        logger.info("Shutdown complete")
-
-
-# Use Click to define CLI with same options as serve command
-@click.command("grpc_server")
-@click.argument("model", type=str)
-@click.option("--tokenizer",
-              type=str,
-              default=None,
-              help="Path | Name of the tokenizer. "
-              "Specify this value only if using TensorRT engine as model.")
-@click.option("--host",
-              type=str,
-              default="0.0.0.0",
-              help="Hostname of the gRPC server.")
-@click.option("--port", type=int, default=50051, help="Port of the gRPC server.")
-@click.option("--backend",
-              type=click.Choice(["pytorch", "trt"]),
-              default="pytorch",
-              help="Set to 'pytorch' for pytorch path. Default is cpp path.")
-@click.option('--log_level',
-              type=click.Choice(severity_map.keys()),
-              default='info',
-              help="The logging level.")
-@click.option("--max_beam_width",
-              type=int,
-              default=BuildConfig.max_beam_width,
-              help="Maximum number of beams for beam search decoding.")
-@click.option("--max_batch_size",
-              type=int,
-              default=BuildConfig.max_batch_size,
-              help="Maximum number of requests that the engine can schedule.")
-@click.option(
-    "--max_num_tokens",
-    type=int,
-    default=BuildConfig.max_num_tokens,
-    help="Maximum number of batched input tokens after padding is removed in each batch."
-)
-@click.option(
-    "--max_seq_len",
-    type=int,
-    default=BuildConfig.max_seq_len,
-    help="Maximum total length of one request, including prompt and outputs. "
-    "If unspecified, the value is deduced from the model config.")
-@click.option("--tp_size", type=int, default=1, help='Tensor parallelism size.')
-@click.option("--pp_size",
-              type=int,
-              default=1,
-              help='Pipeline parallelism size.')
-@click.option("--ep_size",
-              type=int,
-              default=None,
-              help="Expert parallelism size")
-@click.option("--cluster_size",
-              type=int,
-              default=None,
-              help="Expert cluster parallelism size")
-@click.option("--gpus_per_node",
-              type=int,
-              default=None,
-              help="Number of GPUs per node. Default to None, and it will be "
-              "detected automatically.")
-@click.option("--kv_cache_free_gpu_memory_fraction",
-              type=float,
-              default=0.9,
-              help="Free GPU memory fraction reserved for KV Cache, "
-              "after allocating model weights and buffers.")
-@click.option("--trust_remote_code",
-              is_flag=True,
-              default=False,
-              help="Flag for HF transformers.")
-@click.option(
-    "--extra_llm_api_options",
-    type=str,
-    default=None,
-    help="Path to a YAML file that overwrites the parameters specified by trtllm-serve."
-)
-@click.option(
-    "--reasoning_parser",
-    type=click.Choice(list(ReasoningParserFactory.parsers.keys()) + [None]),
-    default=None,
-    help="[Experimental] Specify the parser for reasoning models.",
-)
-@click.option(
-    "--fail_fast_on_attention_window_too_large",
-    is_flag=True,
-    default=False,
-    help="Exit with runtime error when attention window is too large to fit even a single sequence in the KV cache."
-)
-def main(
-    model: str,
-    tokenizer: Optional[str],
-    host: str,
-    port: int,
-    log_level: str,
-    backend: str,
-    max_beam_width: int,
-    max_batch_size: int,
-    max_num_tokens: int,
-    max_seq_len: int,
-    tp_size: int,
-    pp_size: int,
-    ep_size: Optional[int],
-    cluster_size: Optional[int],
-    gpus_per_node: Optional[int],
-    kv_cache_free_gpu_memory_fraction: float,
-    trust_remote_code: bool,
-    extra_llm_api_options: Optional[str],
-    reasoning_parser: Optional[str],
-    fail_fast_on_attention_window_too_large: bool,
-):
-    """
-    TensorRT-LLM gRPC Server
-
-    MODEL: model name | HF checkpoint path | TensorRT engine path
-
-    This server provides a gRPC interface for TensorRT-LLM, designed for
-    high-performance communication with external routers (e.g., sgl-router)
-    using pre-tokenized input and raw token ID output.
-
-    All options are inherited from the standard TensorRT-LLM serve command.
-    """
-    logger.set_level(log_level)
-
-    # Use get_llm_args from serve.py to build LLM arguments
-    # This ensures consistency with the HTTP server
-    llm_args, _ = get_llm_args(
-        model=model,
-        tokenizer=tokenizer,
-        backend=backend,
-        max_beam_width=max_beam_width,
-        max_batch_size=max_batch_size,
-        max_num_tokens=max_num_tokens,
-        max_seq_len=max_seq_len,
-        tensor_parallel_size=tp_size,
-        pipeline_parallel_size=pp_size,
-        moe_expert_parallel_size=ep_size,
-        moe_cluster_parallel_size=cluster_size,
-        gpus_per_node=gpus_per_node,
-        free_gpu_memory_fraction=kv_cache_free_gpu_memory_fraction,
-        num_postprocess_workers=0,  # gRPC doesn't need postprocess workers
-        trust_remote_code=trust_remote_code,
-        reasoning_parser=reasoning_parser,
-        fail_fast_on_attention_window_too_large=fail_fast_on_attention_window_too_large,
-    )
-
-    # Load extra options from YAML if provided
-    if extra_llm_api_options is not None:
-        with open(extra_llm_api_options, 'r') as f:
-            llm_args_extra_dict = yaml.safe_load(f)
-        llm_args = update_llm_args_with_extra_dict(llm_args, llm_args_extra_dict)
-
-    # Run gRPC server
-    asyncio.run(serve_grpc(host, port, llm_args, model))
-
-
-if __name__ == "__main__":
-    main()
